Fix add/change key policies
Add support for *string fields in <Action>Opts structs converted to
cobra command flags.

Convert AddOpts From and Value fields to *string so we can define
policies based on the absence of --value and --from flags.

=> do not error when adding an already existing key if no value is set

Add a "changed" bool state to xconfig.T. Set it to true when changing a
in-memory key. Set it to false on write.

=> don't log "key set" if the value did not actuallty change.