Add the secure keystore "fullpem" and "pkcs" implementation
The pkcs still has issues: the generated p12 is not accepted:
could not parse PKCS12 file, check password,
OpenSSL error error:0308010C:digital envelope routines::unsupported

Also the password is hardcoded for now.

PKCS is discouraged now, so we could drop this command, but it is not clear
what to use instead.