Make "daemon stop" trigger the "maintenance" nmon state
This patch also changes the msgbus.SetNodeMonitor struct, so it
exposes only the fields allowed to change and uses pointers so
the nil value can be the no-change case.

This is cleaner and aligned with the daemonapi.PostNodeMonitor
struct.

Implement the NodeMonitor.State change in the nmon
msgbus.SetNodeMonitor event handler.