Rename the value field of msgbus.*Updated to Value
Trying to use a field name reflecting the field type proved
to be too hard, as copy/paste errors got commited over time.

Use "Value" as a generic name.

This patch also changes the msgbus.SetInstanceMonitor struct, so it
exposes only the fields allowed to change and uses pointers so
the nil value can be the no-change case.

This is cleaner and aligned with the daemonapi.PostInstanceMonitor
struct.