Run network.Setup() on network config change
Add a SectionSig() func to xconfig.T to compute a checksum of
the section key-val pairs.

Add a sig cache to ccfg.ccfg to track network#.* sections
changes on cluster config change events, and to react by executing
network.Setup()

A possible enhancement could be to reconfigure only the nets that
changed instead of all.