Merge node and inst frozen that happened when the daemon was down
Touch a var/last_shutdown file when the nmon routine is canceled.

Compare this last_shutdown mtime with the peer nodes FrozenAt, and
freeze the node if peer was frozen when we were down.

For each object, compare this last_shutdown mtime with the peer
instances FrozenAt, and freeze the local instance if the peer
instance was frozen when we were down.