Avoid Marshal/Unmarshal data in the core/output renderer
This Marshal concrete > Unmarshal in a any-type var was done to
produce a unstructured dataset with keys aligned with the json
tags.

But this method has poor performance issue and the resulting data
has unexpected glitches (for example big int64 value end up in
float-like format, as seen in "pool ls" outputs).

This patch add a unstructed.L type and a Unstructured() map[string]any
type-func to every type passed to the core/output renderer with
tab as the default.

The unstructed type will also allow commands to mangle the
dataset for unit changes and pretty prints, as values can change
type and even new keys can be added.