[util/findmnt] Update commands to avoid nfs hang
We can't use findmnt -J -T {mnt} -S {dev} for nfs because it may hang.
A timeout version of findmnt is not sufficient, we have to differentiate hang but mounted
from not mounted.

So when dev is on nfs, We can't use findmnt -J -T {mnt} -S {dev}
Instead findmnt -J -S {dev} is used, then mnt is filtered within List function.