Honor the "default" cluster name default in rawconfig init()
rawconfig.init() is responsible for setting the clusterName in omcrypto,
which panics when the cluster name is empty.

The current rawconfig.init() code leaves the cluster.name viper key
empty is not defined in the config file nor passed via the env var.

Add a fallback to "default" in this situation.