Fix SetNodeMonitor after a peer that never talked is removed from cluster.nodes
A peer that never talked has an placeholder entry in the nmon.nodeMonitor
cache, that is used to validate the cluster nodes nmon states are ok for
this SetNodeMonitor.

Normally the NodeMonitorDeleted event handler deletes the cache entry,
but if the node is dropped before we receive any hb from it, no
NodeMonitorDeleted is published.

A ForgetPeer event is always published though, so add a cache entry drop
in this event handler too.