Make the client code do the keystore --from to []byte work
This code was in object.Keystore instances. Move it to util/uri
so omcmd and oxcmd can use it.

Remove these from the Keystore interface:

* AddKeyFrom
* PrepareAddKeyFrom
* ChangeKeyFrom
* PrepareChangeKeyFrom

It's much more sane to have the client download or read local
files himself, instead of allowing a read privilege escalation
via the daemon api or "sudo om cfg add --key".