Make task and sync resources optional=false by default
And change the instance status evaluation algo to never merge
the status of these resources into availstatus.

The only role left to optional=true is to continue with the
next resource action on error and don't propagate the error to
the command exitcode.

With this patch, if not explicitely changed by optional=true,
a task run error causes a exitcode 1.