Support task.max_parallel=<int>
Which allows running a task more than once concurrently.

The scheduler still serializes strictly the runs, but the user
can run while the scheduler holds a run too, if max_parallel>1.

The default value is 1, which is the backward compatible
behaviour.

This patch also move common code from restask{host,docker} to
a new restask package.