Fix icfg not restarting after a cfg corruption in the file is corrected
For example, adding a "\n" line in a config file using vi make
icfg exit and the first restart fails.

After fixing the error, the icfg would not get restarted even
if the chmod and write events are seen.

This patch clears the last known config file mtime when the icfg
restart fails, to rearm the icfg start on change event.