Allow nsdev sharing in the ip.netns driver
If 2 ip.netns resource use the same nsdev, assume they use the same mode
and make the 2nd started resource skip the nsdev setup.

On stop, delete the nsdev only if it hosts no ip addrs except the fe80:
ipv6 addr.

This patch permits notably to setup ipv4 and ipv6 addrs on the same
container interface.

Also log the ip commands equivalent to the netlink messages the program
sends.

And add nsdev and netns values to the resource label.