[imon] Introduce "stopped" state in orchestration stop process
This update adds a "stopped" transition state in both the
orchestrateStopped() and stop() functions, creating the flow:
   stopping -> stopped -> idle.

This "stopped" state is utilized in the orchestrateHAStart function to
change the local expectation from started -> none.

We have to remove the local expect started as soon as possible.
Example: automatic stop of flex instances when flex_target is reduced.