Stop announcing syncing and running with POST /.../progress
We want to track this information separately, not to mess
with instance.Monitor.State which is dedicated to exclusive
orchestrations.

The "run" action notably can run in parallel (multiple per task
runs or multiple runs for a single task), so the idle post-action
progress announce is corrupting the instance.Monitor.State.