[imon] Fix missed resource monitoring enable during instance local start
The automatic local instance resource monitoring was enabled during the
detection of local instance status up on idle instance.
But the instance is not always idle (example starting) => the instance
resource monitoring is not enabled as expected and we have to wait for
next instance status evaluation to have instance resource monitoring
enabled.

The automatic local instance resource monitoring is not verified during
update if change calls.

The automatic enable resource monitoring condition is:
  - no orchestration in progress
  - state is idle
  - local expect is not yet started
  - local instance status avail is up