[daemonapi] Remove session id from published label to enable future cache.
This commit removes session ids from labels, to have cacheable label values.

Don't anymore use published label with SessionID or RequesterSessionID.
Labels have a cost, instead move it to Exec... structures so filter is still
possible.

This commit introduces the `RequesterSessionID` field to `Exec`, `ExecFailed`,
and `ExecSuccess` message structures. So it is still possible to filter on data
value.