[imon] Fix: Prevent HA from starting instances in warning state
Issue:
    When a system administrator stops a resource on a high-availability
    (HA) instance with flex topology, the orchestrateHAStart function
    calls orchestrateStarted(), which then starts the instance again.

Previously, orchestrateHAStart would attempt to start the local instance
even if it was in a warning state, which is not the intended behavior.

The recovery of instances with a status of "available warn" is already
managed by resource monitoring when monitor is enabled (monitor is true)
or a restart policy is defined (restart > 0).

When a user stops a resource of an instance, the resource restart and
monitoring are disabled accompanied by the following log message:
  User is stopping some instance resources: disable resource restart
  and monitoring.

Now, orchestrateHAStart no longer attempts to start the local instance
if its status is available warn.