Add support for custom per ip resource dns entries
Expose a "hostname" info key in the ip.netns or ip.cni resources
status.

Let the dns create a <hostname>.<fqdn> record if hostname is
not empty, instead of the default <rindex>.<fqdn> record.

The ip hostname value is the hostname fetched from the resource
pointed by the netns keyword (all container drivers are patched to
implement the GetHostname() interface).

If the dns_name_suffix keyword is set in the ip resource config,
the hostname is appended the dns_name_suffix.

Example:

With ip#1.dns_name_suffix=-backend2 and container#1.hostname=pause:

	pause-backend2.svc1.root.svc.dev2n1.node.reliable-leopard. IN  A6    60    fd01:2345:6789:2900::22
	pause-backend2.svc1.root.svc.reliable-leopard.             IN  A6    60    fd01:2345:6789:2900::22

With ip#1.dns_name_suffix not set and container#1.hostname=pause:

	pause.svc1.root.svc.dev2n1.node.reliable-leopard.          IN  A6    60    fd01:2345:6789:2900::22
	pause.svc1.root.svc.reliable-leopard.                      IN  A6    60    fd01:2345:6789:2900::22

With ip#1.dns_name_suffix not set and container#1.hostname not set:

	1.svc1.root.svc.dev2n1.node.reliable-leopard.              IN  A6    60    fd01:2345:6789:2900::22
	1.svc1.root.svc.reliable-leopard.                          IN  A6    60    fd01:2345:6789:2900::22