Fix the provision orchestration for provisioned=n/a objects
Services like system/svc/vip do not implement the Provision()
method, resulting in their provisioned state being marked as
n/a. Consequently, the orchestration process did not execute
the provision CRM action, leaving both leader and non-leader
instances down. This caused the --wait command to fail, as it
expects the service to be up at the end of the orchestration.

This patch updates the orchestration process to start the leader
instance and initiate the standby mode for non-leader instances
on such services.