Check for context deadline exceeded if cmd.Wait errs
In this case return the context.DeadExceeded to the caller so
he can decide what to do from there.

Before this patch the caller received a sigkill error and could
not differentiate a external cause kill from a timeout.

Update resapp so the timeout is written as a resource warnlog.