Discard non-resolvable dns ipaddrs from the dnsnode ref solver
So the nodes doesn't eval as empty when an admin introduce a
wrong ip in cluster.dns, or when /etc/hosts is missing a node
entry.

This patch also differenciate the case when the config Referrer
reports it does not know the submitted ref, from the case the
ref was recognized but its expansion produced an error.

To achieve this, add a new ErrUnknownReference xconfig error
the Referrer returns, and the caller can test to continue
trying fallback expansions.