Add the volume.install keyword
This new keyword brings finer control over installed files and
directories.

Syntax:

	<path> [from sec|cfg <name> [namespace <ns>]] key <key> [perm <perm>] [user <user>] [group <group>]

Where the namespace default value is the namespace of the declaring object.

Where the perm, user and group default values are defined by the perm, user
and group volume keyword.

If no `from` definition is found, a directory is created as <path>.

When a `from` definition is found, files are created in <path>, with
their content fetched from sec or cfg objects.

Example:

	[volume#1]
	type = shm
	size = 1m
	user = 1001
	group = 1001
	install =
	  /etc/ mode 0750 user 1000 group 1000
	  /etc/ssl/ mode 0700 user 1000 group 1000
	  /etc/ from cfg haproxy namespace test key haproxy.cfg mode 0640 user 1000 group 1000
	  /etc/ssl/front.pem from sec d key fullpem mode 0640 user 1000 group 1000 required
	  /etc/ssl/front.chain from sec d key certificate_chain required
	  /etc/profile.d/ from sec d key etc/profile.d/*
	  /data/