Factorize "mon", "<sel> mon", "daemon status" and "cluster status"
In commoncmd.

This patch changes the args mangling trick in SetExecute():

Now we append {"-s", <sel>} to the args when the command is in
the "om <sel> ..." format.

A cobra persistent flag is added for "-s", so the next parsing
in Execute() will find the CmdFoo.ObjectSelector value from the
flag. MergeSelector() would no longer need the top level <sel>.