Add the config csum to the status.json produced by the CRM
When a container starts and the host executes "instance status -r
-o json" in the container to get the encap service status data,
the opensvc daemon might not be started yet.

In this case the status json does not contain the config csum
that is normally provided by the dameon icfg.

But we need this information to decide if the host's version of
the svc config file must be copied to the container before
status evaluation.

This patch just adds the csum field.