Add the encap service evaluation in actor.resourceStatusEval()
The resource must implement the encaper interface:

	type encaper interface {
	       GetHostname() string
	       GetOsvcRootPath() string
	       EncapCp(context.Context, string, string) error
	       EncapCmd(context.Context, ...string) *exec.Cmd
	}

Execute the "om <path> instance status -r -o json" command with
encaper.EncapCmd().

And store the result in the Encap field of instance.Status.

If the config file is not present (exitcode 2) or the config
csum is different than on the host, push the host svc config
using encaper.EncapCp() and retry the command.

If the result is still unexpected, stop trying.