Fetch keyword definitions from the api in "om node doc"
Instead of using directly the keywords.Store build in the om
executable.

Change the api.KeywordDefinitionItem struct to no longer use
pointers.

Port the keywords.Doc* rendering function to work with
api.KeywordDefinitionItems instead of keyword.Store. Define
these renderers in commoncmd, to share to code with om and ox.