Coalesce imon "status -r" executions
The imon routine executed one status -r command per
InstanceConfigUpdated event.

This patch executes status -r for the first event received if
it is not already running, or remembers it has to run when the
current execution ends.

Use the InstanceStatusUpdated event as the trigger to execute
status -r again if flagged for execution.