Execute the status commands in the background
And prevent orchestrate to continue while a status eval is in
progress.

This patch exposes a new interface from the runner pkg:
runner.Enqueue(prio, errC, fn), to use to start the job in the
background. If errC is a "chan error", the caller can wait
for the fn done via <-errC.

This patch is motivated by the following scenario:

	for i in `seq 20`
	do
		om bug1 set --kw env.random=$RANDOM
	done
	om bug1 abort

Which could end up with a SetInstanceMonitor publish timeout
because the imon select loop is busy doing inline "status -r"
execs.

Open question, shouldn't we apply the same logic to all CRM
command executed by imon ?