Fix completion triggering to early
Wait for a ending whitespace to propose more.

Because if we propose candidates, for example "--live" when
the text is "do switch", the return keypress autocompletes to
"do switch --live" instead of triggering a normal switch.

Worse, when deleting "--live" we get back to the forced
autocompletion choice.

This patch also reorganize the completion code for readability.