[imon] Remove extra instance-status evaluation during startup
omon use both local instance config cache or event to start imon.
So the event is or is not forwarded to imon after imon is started.

imon is started by omon, on instance config updated with 2 != situation:
localhost or peer instance config updated.
When the source event was peer instance config an extra instance status
evalution is called.